requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
numpy==1.26.2
pymongo==4.6.0
pyModbusTCP==0.2.1
ultralytics==8.0.196
python-multipart==0.0.6
websockets==12.0
pyinstaller==6.2.0
python-dotenv
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
Pillow>=9.0.0
PyQt5>=5.15.0
scipy>=1.9.0
torch>=1.12.0
torchvision>=0.13.0
matplotlib>=3.5.0
seaborn>=0.12.0
